backend/requirements.txt
annotated-types==0.7.0
bcrypt==5.0.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
datamatrix==1.0.16
dnspython==2.8.0
et_xmlfile==2.0.0
ezdxf==1.4.2
fastnumbers==5.1.1
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
fonttools==4.60.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
httplib2==0.31.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
json-tricks==3.17.3
MarkupSafe==3.0.3
numpy==2.3.3
openpyxl==3.1.5
prettytable==3.16.0
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.15.1
pyparsing==3.2.5
python-dotenv==1.1.1
qrcodegen==1.8.0
requests==2.32.5
rsa==4.9.1
scipy==1.16.2
segno==1.6.6
svgwrite==1.4.3
tomlkit==0.13.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.14
Werkzeug==3.1.3
